requirements.txt
# requirements.txt
python-dotenv==1.0.1
streamlit==1.36.0
pypdf==4.2.0
python-docx==1.1.0
tqdm==4.66.4
typing-extensions==4.12.2
langchain==0.2.12
langchain-core==0.2.27
langchain-community==0.2.11
langchain-huggingface==0.0.3
huggingface-hub==0.23.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
google-generativeai==0.7.0  # <-- Corrected version
langchain-google-genai==1.0.8


fastapi
uvicorn
python-multipart
pydantic
tiktoken
